Solution to Input Two Floats and Print Their Product

+++ Input Two Floats and Print Their Product@@6788d5012db25efec946f93e/code.c
@@ -1,7 +1,7 @@
 #include <stdio.h>
 int main(){
     float a,b;
-    float product = 1.0;
+    double product = 1.00;
     scanf("%f %f",&a,&b);
     product = a*b;
     printf("Product: %f",product);
